Correct a desynchronization bug when converting songs in RetroTracker
Using a float for a delay value sounds good in theory, but in practice the SPC
side doesn't support fractional delays in this manner, causing the fractional
ticks to get lost instead of compensating for them. Since there's a tick counter
already keeping track of things as a float, I replaced the float versions of the
delay variables with their original integer counterparts and linked them up to
an integer version of the tick counter, which simply grabbed the whole portion
of the float tick counter.