Track what channels are actually being used by the song
Sometimes there's no song data in the track that is converted. If that happens,
then the channel bit is simply disabled so that the track doesn't waste a few
bytes of memory in the process. The local Nchannels report was also modified to
count the number of channels allocated for the song (6 or 8 depending on command
line options) versus the number of channels actually used.