Temporarily force the (relative) instrument volume setting to zero
The way the volume scaler is handled on the SPC side is not quite like a
multiplicative scaler: instead, it implies a minimum note volume, and the note
volume setting offsets off of this. This tends to throw off the way the volume
is scaled on the SPC side, even causing wraparound situations under the wrong
conditions. Thus, this has been de-facto commented out for the time being.