Various optimizations to the SPC code
- Replaced a bunch of AND+CMP operations mostly with BBC/BBS opcodes since it's
otherwise not needed in the accumulator, though in one case (when handling
delay) it was replaced with a bpl/bmi opcode instead since that was in the
accumulator.
- Just before calling play_sample, when zeroing out volume and tmp+3, the
accumulator can be used instead because it was zeroed out.
- At the start of play_sample, the cmp a,#$00 opcode is not needed because the
flag is set at the AND opcode. In addition, pop a doesn't set any flags on the
SPC700 unlike the 65816, therefore we can save a byte by moving when the pop a
opcode is performed.
- In SKS_TRACK_PLAY, cfirst was loaded into the accumulator instead of comparing
it to zero, since it saves a byte.
- Removed a redundant read of play, since its value is already in the
accumulator.
- Prior to calling brr_on, instead of using tmp to store the voice ID, just
directly load it into the X register, since it is not touched.
- In brr_on, directly OR bkof by tmp.